zmq_services/requirements.txt
# Core ZMQ library
pyzmq

# Serialization library
msgpack

# VNPY related libraries (adjust versions as needed based on your existing setup)
# Assuming you have these installed already, but listing them for clarity.
# You might have specific versions pinned in your main project requirements.
# vnpy
# vnpy-ctp

# Added numpy
numpy

# Added for plotting
matplotlib 